requirements.txt
python-dateutil==2.8.2
pandas==1.5.1